fix import, move submodel impl to bayesian msm
